backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
google-generativeai
python-multipart
requests
httpx
python-dotenv
google-cloud-storage
psycopg2-binary
